Added improvements to modal from review
