+++ packages/lib/src/alert/ModalAlertWrapper.tsx
@@ -3,6 +3,7 @@ import styled, { createGlobalStyle } from "styled-components";
 import { responsiveSizes } from "../common/variables";
 import FocusLock from "../utils/FocusLock";
 import { ModalAlertWrapperProps } from "./types";
+import { useEffect } from "react";
 
 const BodyStyle = createGlobalStyle`
   body {
@@ -40,8 +41,22 @@ const ModalAlertContainer = styled.div`
   }
 `;
 
-const ModalAlertWrapper = ({ condition, onClose, children }: ModalAlertWrapperProps) =>
-  condition ? (
+const ModalAlertWrapper = ({ condition, onClose, children }: ModalAlertWrapperProps) => {
+  useEffect(() => {
+    if (condition) {
+      const handleModalAlertKeydown = (event: KeyboardEvent) => {
+        if (event.key === "Escape") {
+          event.preventDefault();
+          onClose?.();
+        }
+      };
+      document.addEventListener("keydown", handleModalAlertKeydown);
+      return () => {
+        document.removeEventListener("keydown", handleModalAlertKeydown);
+      };
+    }
+  }, [condition, onClose]);
+  return condition ? (
     <>
       <BodyStyle />
       {createPortal(
@@ -57,5 +72,6 @@ const ModalAlertWrapper = ({ condition, onClose, children }: ModalAlertWrapperPr
   ) : (
     children
   );
+};
 
 export default ModalAlertWrapper;
